auth_server/pyproject.toml
[project]
name = "oauth2-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "ruff>=0.12.7",
    "sqlmodel>=0.0.24",
]
